// src/app/Models/routefix.ts
export class Routefix {
  id: number;
  seq: number;
  lont: number;
  isWpLeg: boolean;
  distance: number;
  routeId: number;
  wpTime: string;
  directive: string[];
  waypoint: string;
  remarks: string;
  locationMaincords: number;
  locationRadius: number;
  locationAeraGeo: number[];
  locationRef: string;
  locationType: string;
  paxIn: string[];
  paxOut: string[];
  emb: number; //added extra
  deb: number;  //added extra
  cumuldistance: string;
  totalonboard: number;


}